feat: add sepet functionality to FrontendController
This commit adds a new method `sepet` to the `FrontendController` class in `app/Http/Controllers/FrontendController.php`. The `sepet` method accepts a `Request` object and uses `dd()` to dump and die the request data. This functionality is necessary for handling the shopping cart functionality on the frontend.

Also, the `resources/views/layouts/footer.blade.php` file has been modified to include the `layouts.basket` view.

Lastly, the `routes/web.php` file has been updated to include a new route for the `sepet` functionality.