feat: implement sepet functionality in FrontendController
This commit adds the `sepet` method to the `FrontendController` class in `app/Http/Controllers/FrontendController.php`. The `sepet` method accepts a `Request` object and modifies the request data to include quantities for each product. This change is necessary for handling the shopping cart functionality on the frontend.

It also updates the `resources/views/layouts/basket.blade.php` file to remove the quantity controls and includes the `layouts.basket` view in the `resources/views/layouts/footer.blade.php` file.

Lastly, the `routes/web.php` file has been updated to include a new route for the `sepet` functionality.