feat: Update product detail page to use dynamic slug in product name link
This commit updates the `products.blade.php` file to use a dynamic slug in the product name link. The `href` attribute of the `<a>` tag is modified to use the `route()` function with the `page.product.detail` route and the `$product->slug` variable. This change improves the functionality of the website by ensuring that the correct product detail page is linked when the product name is clicked.