feat: Update basket route to use 'basket' instead of 'sepet'
This commit updates the `FrontendController` class in `app/Http/Controllers/FrontendController.php` to change the name of the `sepet` method to `basket`. It also updates the route in `routes/web.php` to use the new method name. This change improves the clarity and consistency of the codebase by using a more descriptive and intuitive name for the basket functionality.