feat: Update category links in index.blade.php and web.php
This commit updates the category links in the index.blade.php file and the web.php file. The links are modified to use the route() function with the 'page.products' route and the $category->slug variable. This change improves the functionality of the website by ensuring that the correct category page is linked when a category name is clicked.